+++ pynecone/__init__.py
@@ -22,7 +22,6 @@ from .event import EVENT_ARG as EVENT_ARG
 from .event import EventChain as EventChain
 from .event import FileUpload as upload_files
 from .event import console_log as console_log
-from .event import get_local_storage as get_local_storage
 from .event import redirect as redirect
 from .event import set_cookie as set_cookie
 from .event import set_focus as set_focus
@@ -39,3 +38,4 @@ from .style import color_mode as color_mode
 from .style import toggle_color_mode as toggle_color_mode
 from .vars import Var as Var
 from .vars import cached_var as cached_var
+from .vars import get_local_storage as get_local_storage
